requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.103.2
h11==0.16.0
idna==3.10
passlib==1.7.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
pypng==0.20220715.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
qrcode==8.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.23.2
